Create coin/math_test.go
